+++ Changelog.md
@@ -1,5 +1,12 @@
 # Releases
 
+## v0.3.5 (??/??/????)
+
+- API breaking changes:
+
+  - move free function: report_quest_env() onto QuestEnv type as
+    `QuestEnv::report_quest_env()`
+
 ## v0.3.4 (29/07/2023)
 
 - Improve documentation